Generality and Usability Enhacements
BioDC is being entirely re-written from scratch for generality and usability. 

Generality: BioDC is now designed to handle any b- or c-type heme in the reduced and oxidized states with any His-[Dist] (where Dist is any proteogenic residue) pair of axial ligands, so long as AMBER force field parameters are available. Parameters for His-His and His-Met b or c-type hemes come with BioDC, more are being developed, and any can be easily added to the data/forcefield directory. As long as the file nomenclature is respected, BioDC will work with your parameters. 

Usability: BioDC has always been an interactive command-line tool. Version 3 keeps to that tradition, improving the command-line interface with menu selections. But version 3 also adds a GUI (so far, only for the Structure Preparation and Relaxation Module). The GUI takes you from PDB download to prepared AMBER topology and coordinate files for a multi-heme cytochrome in minutes. In principle, the program can be used for setting up other structures besides multi-heme cytochromes for molecular dynamics with AMBER.